done with setters
