+++ srcs/parsing/ServerBlockKeymap.cpp
@@ -1,12 +1,12 @@
 
 #include "../../includes/parsing/ServerBlockKeymap.hpp"
-#include "../../includes/parsing/serve.hpp"
+#include "../../includes/parsing/ServerBlock.hpp"
 
 void ServerBlockKeymap::callFunction(const std::string &key, const std::string &value, ServerBlock &instance) const {
     typedef void (ServerBlock::*MemberFuncType)(const std::string&);
     MemberFuncType func = keymap.find(key)->second;
     if (key.size() == 0 || value.size() == 0 || !func) {
-        instance.reportError(ParseException("Invalid keyword or content for keyword in Server Block"));
+        errorHandler::reportError(ParseException("Invalid keyword or content for keyword in Server Block"));
     }
     (instance.*func)(value);
 }
