--- srcs/parsing/cluster.cpp
@@ -1,346 +1,0 @@
-#include "../../includes/parsing/cluster.hpp"
-
-Config::Config() {
-    printf("ConfigFile was created\n");
-}
-
-Config::~Config() {
-}
-
-void Config::reportError(const ParseException &ex) const {
-    throw ex;
-}
-
-// opening config file
-int Config::getTypePath(const std::string &path) {
-    struct stat sb = {};
-    if (stat(path.c_str(), &sb) == 0) {
-        if (S_ISREG(sb.st_mode))
-            return 1;
-        else if (S_ISDIR(sb.st_mode))
-            return 2;
-        else
-            return 3;
-    }
-    return 0;
-}
-
-std::string Config::getContent(const std::string &path) {
-    if (path.empty())
-        return "";
-    std::ifstream file(path.c_str());
-    if (!file.is_open() || !file)
-        return "";
-    std::stringstream buffer;
-    buffer << file.rdbuf();
-    return buffer.str();
-}
-
-bool Config::isFileExistAndReadable(const std::string &path, const std::string &index) {
-    if (getTypePath(index) == 1 && access(index.c_str(), R_OK) == 0)
-        return true;
-    if (getTypePath(path + index) == 1 && access((path + index).c_str(), R_OK) == 0)
-        return true;
-    return false;
-}
-
-const std::string &Config::getFile() const {
-    return conf_file;
-}
-
-const std::vector<ServerBlock> &Config::getServers() const {
-    return servers;
-}
-
-// creating cluster of servers
-// проверить правильно ли работает удаление комментариев
-// проверить правильно ли работает удаление пробелов
-void Config::parseConfig(std::string &content) {
-    // remove comments
-    size_t start_pos = 0;
-    while ((start_pos = content.find('#', start_pos)) != std::string::npos) {
-        size_t end_pos = content.find('\n', start_pos);
-        if (end_pos == std::string::npos)
-            content.erase(start_pos);
-        else
-            content.erase(start_pos, end_pos - start_pos);
-    }
-    // remove whitespaces
-    start_pos = content.find_first_not_of(" \t\n\r\f\v");
-    if (start_pos == std::string::npos) {
-        content.clear();
-        return;
-    }
-    size_t end_pos = content.find_last_not_of(" \t\n\r\f\v");
-    content = content.substr(start_pos, end_pos - start_pos + 1);
-}
-
-
-// Finds the corresponding bracket for the opening bracket at start_bracket_pos
-// Returns the position of the closing bracket or std::string::npos if the closing bracket is not found
-// Arguments: str - string to search in, start_bracket_pos - position of the opening bracket
-size_t findEndingBracket(const std::string &str, size_t start_bracket_pos) {
-    size_t openBracketPos = str.find('{', start_bracket_pos + 1);
-    size_t closeBracketPos = str.find('}', start_bracket_pos + 1);
-    // std::cout << "Open bracket pos: \n\n[" << str.substr(0, openBracketPos) << "]\n\n" << std::endl;
-    // std::cout << "Close bracket pos: \n\n[" << str.substr(0, closeBracketPos) << "]\n\n" << std::endl;
-    if (closeBracketPos == std::string::npos)
-        return std::string::npos;
-    if (openBracketPos == std::string::npos || openBracketPos > closeBracketPos)
-        return closeBracketPos;
-    while (true) {
-        closeBracketPos = findEndingBracket(str, openBracketPos);
-        if (closeBracketPos == std::string::npos)
-            return std::string::npos;
-        openBracketPos = str.find('{', closeBracketPos + 1);
-        closeBracketPos = str.find('}', closeBracketPos + 1);
-        if (openBracketPos == std::string::npos || openBracketPos > closeBracketPos)
-            return closeBracketPos;
-    }
-}
-
-void Config::splitServers(const std::string &content) {
-    size_t current_pos = 0;
-    size_t end_pos = 0;
-    while (end_pos != content.length() - 1) {
-        current_pos = content.find("server {", current_pos);
-        if (current_pos == std::string::npos)
-            break;
-        end_pos = findEndingBracket(content, current_pos + 7);
-        if (end_pos == std::string::npos)
-            reportError(ParseException("Server or location block is not closed"));
-        server_config.push_back(content.substr(current_pos + 8, end_pos - current_pos - 8)); // 8 - length of "server {"
-        current_pos = end_pos + 1;
-    }
-}
-
-std::vector<std::string> Config::splitParams(const std::string &content, const std::string &delim) {
-    std::vector<std::string>	str;
-    std::string::size_type		start, end;
-
-    start = end = 0;
-    while (start != std::string::npos)
-    {
-        end = content.find_first_of(delim, start);
-        if (end == std::string::npos)
-            break;
-        std::string tmp = content.substr(start, end - start);
-        str.push_back(tmp);
-        start = content.find_first_not_of(delim, end);
-    }
-    return (str);
-}
-
-std::vector<std::string> Config::SplitAndCutLocations(std::string &config) const{
-    size_t current_pos = 0;
-    size_t end_pos = 0;
-    std::vector<std::string> locations;
-
-    while (current_pos != std::string::npos) {
-        current_pos = config.find("location");
-
-        if (current_pos == std::string::npos)
-            break;
-        size_t opening_bracket_pos = config.find('{', current_pos);
-        if (opening_bracket_pos == std::string::npos)
-            reportError(ParseException("Location block is not opened"));
-        end_pos = findEndingBracket(config, opening_bracket_pos + 1);
-        if (end_pos == std::string::npos)
-            reportError(ParseException("Server or location block is not closed"));
-        locations.push_back(config.substr(current_pos + 8, end_pos - opening_bracket_pos - 1));
-        config = config.substr(0, current_pos) + config.substr(end_pos + 1);
-        current_pos = end_pos + 1;
-    }
-    return locations;
-}
-
-std::string trimWhitespace(std::string &str) {
-    size_t start = str.find_first_not_of(" \t\n\r\f\v");
-    if (start == std::string::npos)
-        return "";
-    size_t end = str.find_last_not_of(" \t\n\r\f\v");
-    return str.substr(start, end - start + 1);
-}
-
-LocationBlock Config::CreateLocation(std::string &locationTxt) const {
-    LocationBlock location;
-    std::string path = locationTxt.substr(0, locationTxt.find('{'));
-    location.setPath(path);
-    locationTxt = locationTxt.substr(locationTxt.find('{') + 1);
-    std::vector<std::string> params = splitParams(locationTxt, ";");
-    for (size_t i = 0; i < params.size(); i++) {
-        params[i] = trimWhitespace(params[i]);
-        std::string key = params[i].substr(0, params[i].find(' '));
-        std::string value = params[i].substr(params[i].find(' ') + 1);
-        location.getKeymap().callFunction(key, value, location);
-    }
-    return location;
-}
-
-std::vector<LocationBlock> Config::ParseConfigLocations(std::string &config) const {
-    std::vector<std::string> locationsTxt = SplitAndCutLocations(config);
-    std::vector<LocationBlock> locationBlocks;
-    for (size_t i = 0; i < locationsTxt.size(); i++) {
-        locationBlocks.push_back(CreateLocation(locationsTxt[i]));
-    }
-    return std::vector<LocationBlock>();
-}
-
-// этот метод - полный костыль
-// нужно будет переделать, чтобы он правильно парсил параметры сервера
-ServerBlock Config::createServer(std::string &config) const {
-    ServerBlock server;
-    std::vector<std::string> params;
-    std::vector<std::string> error_codes;
-    // bool loca_flag = false;
-    // bool size_flag = false;
-    // bool index_flag = false;
-
-    server.setLocations(ParseConfigLocations(config));
-    params = splitParams(config, std::string(";"));
-    for (size_t i = 0; i < params.size(); ++i) {
-        printf("Param %lu: '%s'\n", i, params[i].c_str());
-    }
-
-    for (size_t i = 0; i < params.size(); i++) {
-        params[i] = trimWhitespace(params[i]);
-        std::string key = params[i].substr(0, params[i].find(' '));
-        std::string value = params[i].substr(params[i].find(' ') + 1);
-        server.getKeymap().callFunction(key, value, server);
-    }
-
-//     if (params.size() < 3)
-//         reportError(ParseException("Server block is empty"));
-//     for (size_t i = 0; i < params.size(); ++i) {
-//         if (params[i] == "listen" && (i + 1) < params.size() && !loca_flag) {
-//             if (server.getPort()) {
-//                 reportError(ParseException("Duplicate Port"));
-//             }
-//             server.setPort(params[++i]);
-//         }
-//         else if (params[i] == "host" && (i + 1) < params.size() && !loca_flag) {
-//             if (server.getHost())
-//                 reportError(ParseException("Duplicate Host"));
-//             server.setHost(params[++i]);
-//         }
-//         else if (params[i] == "root" && (i + 1) < params.size() && !loca_flag) {
-//             if (!server.getRoot().empty())
-//                 reportError(ParseException("Duplicate Root"));
-//             server.setRoot(params[++i]);
-//         }
-//         else if (params[i] == "server_name" && (i + 1) < params.size() && !loca_flag) {
-//             if (!server.getServerName().empty())
-//                 reportError(ParseException("Duplicate Server name"));
-//             server.setServerName(params[++i]);
-//         }
-//         else if (params[i] == "index" && (i + 1) < params.size() && !loca_flag) {
-//             if (!server.getIndex().empty())
-//                 reportError(ParseException("Duplicate Index"));
-//             server.setIndex(params[++i]);
-//         }
-//         else if (params[i] == "client_max_body_size" && (i + 1) < params.size() && !loca_flag) {
-//             if (size_flag)
-//                 reportError(ParseException("Duplicate Client max body size"));
-//             server.setClientMaxBodySize(params[i + 1]);
-//             size_flag = true;
-//         }
-//         else if (params[i] == "autoindex" && (i + 1) < params.size() && !loca_flag) {
-//             if (index_flag)
-//                 reportError(ParseException("Duplicate Autoindex"));
-//             server.setAutoindex(params[++i]);
-//             index_flag = true;
-//         }
-//         else if (params[i] == "error_page" && (i + 1) < params.size() && !loca_flag) {
-//             ; // тут прописать логику для создания вектора error_pages
-//         }
-//         else if (params[i] == "location" && (i + 1) < params.size()) {
-//             // создать location
-//             loca_flag = true;
-//         }
-// //        else if (params[i] != "}" && params[i] != "{") {
-// //            if (loca_flag)
-// //                reportError(ParseException("Location block is not closed or param after location"));
-// //            reportError(ParseException("'" + params[i] + "' invalid parameter"));
-// //        }
-//     }
-//     if (server.getRoot().empty())
-//         server.setRoot("/;");
-//     if (server.getHost() == 0)
-//         server.setHost("localhost;");
-//     if (server.getIndex().empty())
-//         server.setIndex("index.html;");
-    // check if index file exists and readable
-    // check if location is not duplicated
-    // check if errorpages are reachable and readable
-    return ServerBlock();
-}
-
-void Config::createCluster(const std::string &config_file) {
-    // проверка корректности файла
-    conf_file = config_file;
-    std::string content = getContent(config_file);
-
-    if (content.empty())
-        reportError(ParseException(config_file + " is empty"));
-    if (access(config_file.c_str(), R_OK) != 0)
-        reportError(ParseException("Can't read file " + config_file));
-    if (getTypePath(config_file) != 1)
-        reportError(ParseException("File " + config_file + " is not a regular file"));
-
-    // парсинг конфига
-    parseConfig(content);
-    splitServers(content);
-
-    // создание серверов и готово кластера
-    for (size_t i = 0; i < server_config.size(); ++i) {
-        ServerBlock server = createServer(server_config[i]);
-        servers.push_back(server);
-    }
-    printf("Cluster was created\n");
-}
-
-void Config::print() const {
-    std::cout << "------------- Config -------------" << std::endl;
-    for (size_t i = 0; i < servers.size(); i++)
-    {
-        std::cout << "Server #" << i + 1 << std::endl;
-        std::cout << "Server name: " << servers[i].getServerName() << std::endl;
-        std::cout << "Host: " << servers[i].getHost() << std::endl;
-        std::cout << "Root: " << servers[i].getRoot() << std::endl;
-        std::cout << "Index: " << servers[i].getIndex() << std::endl;
-        std::cout << "Port: " << servers[i].getPort() << std::endl;
-        std::cout << "Max BSize: " << servers[i].getClientMaxBodySize() << std::endl;
-        std::cout << "Error pages: " << servers[i].getErrorPages().size() << std::endl;
-//        std::map<short, std::string>::const_iterator it = servers[i].getErrorPages().begin();
-//        while (it != servers[i].getErrorPages().end())
-//        {
-//            std::cout << (*it).first << " - " << it->second << std::endl;
-//            ++it;
-//        }
-        std::cout << "Locations: " << servers[i].getLocations().size() << std::endl;
-        std::vector<LocationBlock>::const_iterator itl = servers[i].getLocations().begin();
-        while (itl != servers[i].getLocations().end())
-        {
-            std::cout << "name location: " << itl->getPath() << std::endl;
-//            std::cout << "methods: " << itl->getPrintMethods() << std::endl;
-//            std::cout << "index: " << itl->getIndexLocation() << std::endl;
-            if (itl->getCgiPath().empty())
-            {
-//                std::cout << "root: " << itl->getRootLocation() << std::endl;
-                if (!itl->getReturn().empty())
-                    std::cout << "return: " << itl->getReturn() << std::endl;
-                if (!itl->getAlias().empty())
-                    std::cout << "alias: " << itl->getAlias() << std::endl;
-            }
-            else
-            {
-//                std::cout << "cgi root: " << itl->getRootLocation() << std::endl;
-                std::cout << "sgi_path: " << itl->getCgiPath().size() << std::endl;
-//                std::cout << "sgi_ext: " << itl->getCgiExtension().size() << std::endl;
-            }
-            ++itl;
-        }
-        itl = servers[i].getLocations().begin();
-        std::cout << "-----------------------------" << std::endl;
-    }
-}
