+++ test.conf
@@ -4,7 +4,6 @@ server {
     client_max_body_size 1024;
     server_name localhost;
 
-
     location / {
         client_max_body_size 1023;
         root /www/images/;
